+++ SEQTaRget/R/SEQexpand.R
@@ -29,7 +29,7 @@ SEQexpand <- function(params) {
                        params@visit.denominator, params@visit.numerator,
                        params@deviation.col, tx_bas)
     }
-    vars <- unique(c(unlist(strsplit(vars.intake, "\\+|\\*|\\:")),
+    vars <- unique(c(formula_vars(vars.intake),
                      params@treatment, params@cense, params@cense.eligible, params@visit,
                      params@compevent, unlist(params@weight.eligible_cols), params@subgroup))
     vars.nin <- c("dose", "dose_sq", params@time, paste0(params@time, params@indicator.squared), "tx_lag", "censored")